third_party/WebKit/LayoutTests/fast/dom/navigator-vibration.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test for the Vibration API.');

shouldBeTrue("navigator.vibrate(0);");
shouldBeTrue("navigator.vibrate([]);");
shouldBeTrue("navigator.vibrate(1);");
shouldBeTrue("navigator.vibrate([1, 2]);");
shouldBeTrue("navigator.vibrate([1, 2, 3]);");

// Max pattern entry value.
shouldBeTrue("navigator.vibrate(10000);");

// Larger than max but the value is truncated and execution proceeds.
shouldBeTrue("navigator.vibrate(10001);");

shouldBeTrue("navigator.vibrate(-1);");

// Maximum pattern length.
shouldBeTrue("var p = []; for (var i = 0; i < 99; i++) p[i] = 1; navigator.vibrate(p);");

// Larger than max but the pattern is truncated and execution proceeds.
shouldBeTrue("var p = []; for (var i = 0; i < 100; i++) p[i] = 1; navigator.vibrate(p);");
shouldThrow("navigator.vibrate();");
shouldThrow("navigator.vibrate([1], [2]);");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
